requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
mysql-connector-python==8.4.0
python-dotenv==1.1.1